% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/222_export_opt.r
\name{Initial Control Functions}
\alias{.zmqopt_init}
\alias{Initial Control Functions}
\title{Initial controls in pbdZMQ}
\usage{
.zmqopt_init(envir = .GlobalEnv)
}
\arguments{
\item{envir}{an environment where ZMQ controls locate}
}
\value{
\code{.zmqopt_init()} initial the ZMQ control
at \code{envir}.
}
\description{
Initial control functions
}
\details{
\code{.zmqopt_init()} initials default ZMQ controls.
}
\examples{
\dontrun{
library(pbdZMQ, quietly = TRUE)

ls(.pbd_env)
rm(.pbd_env)
.zmqopt_init()
ls(.pbd_env)

.pbd_env$ZMQ.SR$BLOCK
.pbd_opt(bytext = "ZMQ.SR$BLOCK = 0L")
}

}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\references{
ZeroMQ/4.1.0 API Reference:
\url{http://api.zeromq.org/4-1:_start}

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\seealso{
\code{\link{.pbd_env}}.
}
\keyword{programming}

