% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.r
\name{address}
\alias{address}
\title{Form an Address/Endpoint}
\usage{
address(host, port, transport = "tcp")
}
\arguments{
\item{host}{The host ip address or url.}

\item{port}{A port; necessary for all transports except ipc.}

\item{transport}{The transport protocol.  Choices are "inproc", "ipc", "tcp", and
"pgm"/"epgm" for local in-process (inter-thread), local 
inter-process, tcp, and pgm, respectively.}
}
\value{
An address, for use with pbdZMQ functions.
}
\description{
A notationally convenient function for forming addresses/endpoints.
It's a simple wrapper around the \code{paste0()} function.
}
\examples{
address("localhost", 55555)

}
\seealso{
\code{\link{zmq.bind}}
}
\author{
Drew Schmidt
}
