% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_zmq_transfer.r
\name{File Transfer Functions}
\alias{File Transfer Functions}
\alias{zmq.recvfile}
\alias{zmq.sendfile}
\title{File Transfer Functions}
\usage{
zmq.sendfile(socket, filename, flags = .pbdZMQEnv$ZMQ.SR$BLOCK)

zmq.recvfile(socket, filename, flags = .pbdZMQEnv$ZMQ.SR$BLOCK)
}
\arguments{
\item{socket}{a ZMQ socket}

\item{filename}{the name of the file to send}

\item{flags}{a flag for the method used by \code{zmq_sendfile} and
\code{zmq_recvfile}}
}
\value{
\code{zmq.sendfile()} and \code{zmq.recvfile()} return
number of bytes (invisible) in the sent message if successful,
otherwise returns -1 (invisible) and sets \code{errno} to the error
value, see ZeroMQ manual for details.
}
\description{
File transfer functions
}
\details{
\code{zmq.sendfile()} is a high level R function calling ZMQ C API
\code{zmq_send()} sending the data stored at \code{filename} out in chunks.

\code{zmq.recvfile()} is a high level R function calling ZMQ C API
\code{zmq_recv()} receiving data from \code{zmq.sendfile()} and
storing it in \code{filename}.

\code{buf.type} currently supports \code{char} and \code{raw} which are both
in R object format.
}
\examples{
\dontrun{
### Using request-reply pattern.

### At the server, run next in background or the other window.
library(pbdZMQ, quietly = TRUE)

context <- zmq.ctx.new()
responder <- zmq.socket(context, .pbdZMQEnv$ZMQ.ST$REP)
zmq.bind(responder, "tcp://*:5555")
for(i.res in 1:5){
  buf <- zmq.recv(responder, 10L)
  cat(buf$buf, "\\n")
  Sys.sleep(0.5)
  zmq.send(responder, "World")
}
zmq.close(responder)
zmq.ctx.destroy(context)


### At a client, run next in foreground.
library(pbdZMQ, quietly = TRUE)

context <- zmq.ctx.new()
requester <- zmq.socket(context, .pbdZMQEnv$ZMQ.ST$REQ)
zmq.connect(requester, "tcp://localhost:5555")
for(i.req in 1:5){
  cat("Sending Hello ", i.req, "\\n")
  zmq.send(requester, "Hello")
  buf <- zmq.recv(requester, 10L)
  cat("Received World ", i.req, "\\n")
}
zmq.close(requester)
zmq.ctx.destroy(context)
}

}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\references{
ZeroMQ/4.1.0 API Reference:
\url{http://api.zeromq.org/4-1:_start}

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\seealso{
\code{\link{zmq.msg.send}()}, \code{\link{zmq.msg.recv}()}.
}
\keyword{programming}

