% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyzmq_wrapper.r
\docType{class}
\name{PyZMQ-like Interface}
\alias{PyZMQ-like Interface}
\alias{pyzmq}
\alias{Socket}
\alias{Context}
\alias{zmq}
\title{R6 ZMQ Interface}
\format{An \code{\link{R6Class}} generator object}
\usage{
Socket

Context

zmq
}
\description{
R6 ZMQ Interface
}
\examples{
\dontrun{
context = zmq$Context()
socket = context$socket("ZMQ_REQ")
socket$connect("tcp://localhost:5555")
### etc...
}

}
\references{
Modeled after the PyZMQ interface \url{https://zeromq.github.io/pyzmq/api/zmq.html}
}
\author{
Drew Schmidt
}
\keyword{data}
\keyword{datasets}
