% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rzmq_wrapper.r
\name{Wrapper Functions for rzmq}
\alias{Wrapper Functions for rzmq}
\alias{bind.socket}
\alias{connect.socket}
\alias{init.context}
\alias{init.socket}
\alias{receive.socket}
\alias{send.socket}
\title{All Wrapper Functions for rzmq}
\usage{
send.socket(socket, data, send.more = FALSE, serialize = TRUE)

receive.socket(socket, unserialize = TRUE, dont.wait = FALSE)

init.context()

init.socket(context, socket.type)

bind.socket(socket, address)

connect.socket(socket, address)
}
\arguments{
\item{socket}{A ZMQ socket.}

\item{data}{An R object.}

\item{send.more}{Logical; will more messages be sent?}

\item{serialize,unserialize}{Logical; determines if serialize/unserialize should be called
on the sent/received data.}

\item{dont.wait}{Logical; determines if reception is blocking.}

\item{context}{A ZMQ context.}

\item{socket.type}{The type of ZMQ socket.}

\item{address}{A valid address.  See details.}
}
\value{
TODO
}
\description{
Wrapper functions for backwards compatibility with rzmq.
}
\details{
TODO
}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\references{
ZeroMQ/4.1.0 API Reference:
\url{http://api.zeromq.org/4-1:_start}

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\keyword{internal}

