\name{Wrapper Functions for rzmq}
\alias{send.socket}
\alias{receive.socket}
\alias{init.context}
\alias{init.socket}
\alias{bind.socket}
\alias{connect.socket}
\title{ All Wrapper Functions for rzmq }
\description{ All wrapper functions for rzmq }
\references{
  ZeroMQ/4.1.0 API Reference:
  \url{http://api.zeromq.org/4-1:_start}

  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\keyword{internal}
