% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_pkgso_rpath.r
\name{Overwrite rpath}
\alias{Overwrite rpath}
\alias{overwrite.pkgso.rpath}
\alias{get.pkgso.rpath}
\title{Overwrite and get shared library rpath in osx}
\usage{
overwrite.pkgso.rpath(package, targetpkg)

get.pkgso.rpath(package, verbose = TRUE)
}
\arguments{
\item{package}{for where \code{package.so} will be checked or updated}

\item{targetpkg}{for target pkg where \code{lib*.*.dylib} is located}

\item{verbose}{if printing messages}
}
\description{
Overwrite and get shared library rpath (e.g. \code{lib*.*.dylib}) in osx.
Typically, it is called by \code{.onLoad()} to update rpath if a
\code{lib*.*.dylib} was installed in a personal directory. Permission may
be needed (e.g. \code{sudo}) to overwrite the shared libraries.
}
\details{
\code{overwrite.pkgso.rpath()} update rpath of \code{package.so} for whom is
linked with \code{lib*.*.dylib} shipped with targetpkg in osx.

\code{get.pkgso.rpath()} get rpath for \code{package.so} in osx.
}
\examples{
\dontrun{
get.pkgso.rpath("JuniperKernel")
}

}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\keyword{compile}
