\name{ZMQ Control Functions}
\alias{ZMQ.ST}
\alias{ZMQ.SO}
\alias{ZMQ.SR}
\alias{ZMQ.MC}

\title{Sets of controls in pbdZMQ.}
\description{
  These sets of controls are used to provide default values in this package.
}
\seealso{
  \code{\link{.pbdZMQEnv}}.
}
\references{
  ZeroMQ/4.1.0 API Reference:
  \url{http://api.zeromq.org/4-1:_start}

  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\keyword{global variables}
