% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_zmq.r
\name{ZMQ Flags}
\alias{ZMQ Flags}
\alias{get.zmq.ldflags}
\alias{get.zmq.cppflags}
\title{ZMQ Flags}
\usage{
get.zmq.ldflags(arch = "", package = "pbdZMQ")

get.zmq.cppflags(arch = "", package = "pbdZMQ")
}
\arguments{
\item{arch}{'' (default) for non-windows or '/i386' and '/ix64' for windows}

\item{package}{the pbdZMQ package}
}
\value{
flags to compile and link with ZMQ.
}
\description{
ZMQ Flags
}
\details{
\code{get.zmq.cppflags()} gets CFLAGS or CPPFLAGS

\code{get.zmq.ldflags()} gets LDFLAGS
}
\examples{
\dontrun{
get.zmq.cppflags(arch = '/i386')
get.zmq.ldflags(arch = '/x64')
}

}
\references{
ZeroMQ/4.1.0 API Reference:
\url{http://api.zeromq.org/4-1:_start}

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\keyword{compile}
