\name{pbdZMQ Internal Functions}
\alias{zmq.msg.init}
\alias{zmq.msg.close}
\alias{zmq.send.char}
\alias{zmq.send.raw}
\alias{zmq.recv.char}
\alias{zmq.recv.raw}
\title{ All pbdZMQ Internal Functions }
\description{ All pbdZMQ internal functions }
\references{
  ZeroMQ/4.1.0 API Reference:
  \url{http://api.zeromq.org/4-1:_start}

  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\keyword{internal}
