% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_pkgso_rpath.r
\name{Overwrite rpath}
\alias{Overwrite rpath}
\alias{overwrite.pkgso.rpath}
\alias{overwrite.dylib.id}
\alias{get.pkgso.rpath}
\alias{get.pkgso.id}
\alias{get.alldylib.rpath}
\alias{get.alldylib.id}
\title{Overwrite and get shared library rpath and id in osx}
\usage{
overwrite.pkgso.rpath(package, targetpkg)

overwrite.dylib.id(package, targetpkg)

get.pkgso.rpath(package, verbose = TRUE)

get.pkgso.id(package, verbose = TRUE)

get.alldylib.rpath(package, verbose = TRUE)

get.alldylib.id(package, verbose = TRUE)
}
\arguments{
\item{package}{for where \code{package.so} will be checked or updated}

\item{targetpkg}{for target pkg where \code{lib*.*.dylib} is located}

\item{verbose}{if printing messages}
}
\description{
Overwrite and get shared library rpath (e.g. \code{lib*.*.dylib}) in osx.
Typically, it is called by \code{.onLoad()} to update rpath if a
\code{lib*.*.dylib} was installed in a personal directory. Permission may
be needed (e.g. \code{sudo}) to overwrite the shared libraries.
All functions here are for osx only. The commands \code{otool} and
\code{install_name_tool} are requried as well.
}
\details{
\code{overwrite.pkgso.rpath()} update rpath of \code{package.so} for whom is
linked with \code{lib*.dylib} shipped with targetpkg.

\code{overwrite.dylib.id()} update id of \code{package/libs/libpkg.dylib}
where \code{libpkg.dylib} is specified.

\code{get.pkgso.rpath()} get rpath of \code{package.so}.

\code{get.pkgso.id()} get id of \code{package.so}.

\code{get.alldylib.rpath()} get rpath of all \code{package/libs/lib*.dylib}.

\code{get.alldylib.id()} get id of all \code{package/libs/lib*.dylib}.
}
\examples{
\dontrun{
get.pkgso.rpath("JuniperKernel")
}

}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\keyword{compile}
