% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_zmq.r
\name{ZMQ Flags}
\alias{ZMQ Flags}
\alias{get.zmq.ldflags}
\alias{get.zmq.cppflags}
\alias{test.load.zmq}
\alias{get.pbdZMQ.ldflags}
\title{ZMQ Flags}
\usage{
get.zmq.ldflags(arch = "", package = "pbdZMQ")

get.zmq.cppflags(arch = "", package = "pbdZMQ")

test.load.zmq(arch = "", package = "pbdZMQ")

get.pbdZMQ.ldflags(arch = "", package = "pbdZMQ")
}
\arguments{
\item{arch}{'' (default) for non-windows or '/i386' and '/ix64' for windows}

\item{package}{the pbdZMQ package}
}
\value{
flags to compile and link with ZMQ.
}
\description{
ZMQ Flags
}
\details{
\code{get.zmq.cppflags()} gets CFLAGS or CPPFLAGS

\code{get.zmq.ldflags()} gets LDFLAGS for libzmq.so, libzmq.dll, or libzmq.*.dylib

\code{get.pbdZMQ.ldflags()} gets LDFLAGS for pbdZMQ.so or pbdZMQ.dll

\code{test.load.zmq()} tests load libzmq and pbdZMQ shared libraries
}
\examples{
\dontrun{
get.zmq.cppflags(arch = '/i386')
get.zmq.ldflags(arch = '/x64')
get.pbdZMQ.ldflags(arch = '/x64')
test.load.zmq(arch = '/x64')
}

}
\references{
ZeroMQ/4.1.0 API Reference:
\url{https://libzmq.readthedocs.io/en/zeromq4-1/}

Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\keyword{compile}
